requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=9.5.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
faiss-cpu>=1.7.4
streamlit>=1.28.0
opencv-python>=4.8.0
tqdm>=4.65.0
pandas>=2.0.0
seaborn>=0.12.0
kaggle>=1.5.16
gdown>=4.7.1